+++ src/ansic/difflist.h
@@ -12,7 +12,7 @@
 
 //struct diff_item_s
 //{
-//    int n;		// line number
+//    long n;		// line number
 //    const char *line;	// diff string
 //};
 
@@ -25,7 +25,7 @@ struct diff_list_s
 struct diff_iterator
 {
     TAILQ_ENTRY(diff_iterator) entries;		/* Linked list prev./next entry */
-    int n;		// line number
+    long n;		// line number
     char *line;	// diff string
 };
 
@@ -33,10 +33,10 @@ struct diff_iterator
 
 struct diff_list_s *diff_new(void);
 void diff_delete(struct diff_list_s *list);
-void diff_add_line(struct diff_list_s *list, int n, char *line);
-void diff_add_line_i(struct diff_iterator **iterator, int n, char *line);
-void diff_remove_line(struct diff_list_s *list, int n);
-void diff_remove_line_i(struct diff_iterator **iterator, int n);
+void diff_add_line(struct diff_list_s *list, long n, char *line);
+void diff_add_line_i(struct diff_iterator **iterator, long n, char *line);
+void diff_remove_line(struct diff_list_s *list, long n);
+void diff_remove_line_i(struct diff_iterator **iterator, long n);
 struct diff_iterator *diff_get_first(struct diff_list_s *list);
 struct diff_iterator *diff_get_last(struct diff_list_s *list);
 struct diff_iterator *diff_get_current(struct diff_list_s *list);
@@ -44,12 +44,12 @@ void diff_next(struct diff_list_s *list);
 void diff_previous(struct diff_list_s *list);
 void diff_iterator_next(struct diff_iterator **iterator);
 void diff_iterator_previous(struct diff_iterator **iterator);
-struct diff_iterator *diff_go_before_line(struct diff_list_s *list, int n);
-struct diff_iterator *diff_go_after_line(struct diff_list_s *list, int n);
-void diff_iterator_go_before_line(struct diff_iterator **iterator, int n);
-void diff_iterator_go_after_line(struct diff_iterator **iterator, int n);
+struct diff_iterator *diff_go_before_line(struct diff_list_s *list, long n);
+struct diff_iterator *diff_go_after_line(struct diff_list_s *list, long n);
+void diff_iterator_go_before_line(struct diff_iterator **iterator, long n);
+void diff_iterator_go_after_line(struct diff_iterator **iterator, long n);
 
 const char *diff_get_line(struct diff_iterator *iterator);
-int diff_get_line_nr(struct diff_iterator *iterator);
+long diff_get_line_nr(struct diff_iterator *iterator);
 
 #endif /* SRC_ANSIC_DIFFLIST_H_ */
